refactor: add v2 folder everywhere
