+++ jest.config.packages.js
@@ -86,6 +86,12 @@ module.exports = {
   // Here we ensure that Jest resolves `@metamask/*` imports to the uncompiled source code for packages that live in this repo.
   // NOTE: This must be synchronized with the `paths` option in `tsconfig.packages.json`.
   moduleNameMapper: {
+    '^@metamask/(.+)/v2$': [
+      '<rootDir>/../$1/src/v2',
+      // Some @metamask/* packages we are referencing aren't in this monorepo,
+      // so in that case use their published versions
+      '<rootDir>/../../node_modules/@metamask/$1/v2',
+    ],
     '^@metamask/(.+)$': [
       '<rootDir>/../$1/src',
       // Some @metamask/* packages we are referencing aren't in this monorepo,
