+++ packages/keyring-eth-hd/src/index.ts
@@ -5,4 +5,3 @@ export type {
   HDKeyringOptions,
   HDKeyringAccountSelectionOptions,
 } from './hd-keyring';
-export { HdKeyringV2, type HdKeyringV2Options } from './hd-keyring-v2';
