+++ packages/keyring-eth-hd/src/v2/index.ts
@@ -1,0 +1,1 @@
+export { HdKeyringV2, type HdKeyringV2Options } from './hd-keyring';
