+++ packages/keyring-eth-ledger-bridge/src/index.ts
@@ -1,5 +1,4 @@
 export * from './ledger-keyring';
-export * from './ledger-keyring-v2';
 export * from './ledger-iframe-bridge';
 export * from './ledger-mobile-bridge';
 export type * from './ledger-bridge';
