+++ packages/keyring-eth-ledger-bridge/src/v2/index.ts
@@ -1,0 +1,1 @@
+export * from './ledger-keyring';
