+++ packages/keyring-eth-qr/src/index.ts
@@ -17,8 +17,3 @@ export {
   type SerializedQrKeyringState,
   type SerializedUR,
 } from './qr-keyring';
-export {
-  QrKeyringV2,
-  type QrKeyringV2Options,
-  type QrAccountModeCreateOptions,
-} from './qr-keyring-v2';
