+++ packages/keyring-eth-qr/src/v2/index.ts
@@ -1,0 +1,5 @@
+export {
+  QrKeyringV2,
+  type QrKeyringV2Options,
+  type QrAccountModeCreateOptions,
+} from './qr-keyring';
