+++ packages/keyring-eth-simple/src/index.ts
@@ -1,5 +1,1 @@
 export { default } from './simple-keyring';
-export {
-  SimpleKeyringV2,
-  type SimpleKeyringV2Options,
-} from './simple-keyring-v2';
