+++ packages/keyring-eth-simple/src/v2/index.ts
@@ -1,0 +1,1 @@
+export { SimpleKeyringV2, type SimpleKeyringV2Options } from './simple-keyring';
