+++ packages/keyring-eth-trezor/src/v2/index.ts
@@ -1,0 +1,2 @@
+export * from './trezor-keyring';
+export * from './onekey-keyring';
