+++ packages/keyring-eth-trezor/src/v2/trezor-keyring.test.ts
@@ -5,18 +5,18 @@ import {
   EthScope,
   type KeyringAccount,
   KeyringAccountEntropyTypeOption,
-  KeyringType,
 } from '@metamask/keyring-api';
+import { KeyringType } from '@metamask/keyring-api/v2';
 import HDKey from 'hdkey';
 
-import type { TrezorBridge } from './trezor-bridge';
-import { TrezorKeyring } from './trezor-keyring';
 import {
   BIP44_HD_PATH_PREFIX,
   LEGACY_MEW_PATH_PREFIX,
   SLIP0044_TESTNET_PATH_PREFIX,
   TrezorKeyringV2,
-} from './trezor-keyring-v2';
+} from './trezor-keyring';
+import type { TrezorBridge } from '../trezor-bridge';
+import { TrezorKeyring } from '../trezor-keyring';
 
 /**
  * Type alias for Trezor keyring accounts (always BIP-44 derived).
