+++ packages/keyring-snap-bridge/src/SnapKeyring.ts
@@ -77,7 +77,6 @@ import type {
   SnapKeyringMessenger,
 } from './SnapKeyringMessenger';
 import { SNAP_KEYRING_NAME } from './SnapKeyringMessenger';
-import { SnapKeyringV2 } from './SnapKeyringV2';
 import type { SnapMessage } from './types';
 import { SnapMessageStruct } from './types';
 import {
@@ -88,6 +87,7 @@ import {
   toJson,
   unique,
 } from './util';
+import { SnapKeyringV2 } from './v2/SnapKeyring';
 
 export const SNAP_KEYRING_TYPE = 'Snap Keyring';
