+++ packages/keyring-snap-bridge/src/index.ts
@@ -1,5 +1,4 @@
 export * from './options';
 export * from './types';
 export * from './SnapKeyring';
-export * from './SnapKeyringV2';
 export type * from './SnapKeyringMessenger';
